Move auxflash loading to HubrisArchive
